feat(Sections.java): Sections 일급컬렉션 구현
- line의 addSection 로직을 Sections 로 이동
