+++ src/main/java/nextstep/subway/line/application/LineService.java
@@ -58,7 +58,7 @@ public class LineService {
         Line line = findLineById(lineId);
         Station upStation = stationService.findStationById(request.getUpStationId());
         Station downStation = stationService.findStationById(request.getDownStationId());
-        line.addSectionV2(new Section(line, upStation, downStation, request.getDistance()));
+        line.addSection(new Section(line, upStation, downStation, request.getDistance()));
     }
 
     public void removeLineStation(Long lineId, Long stationId) {
